This approach uses `pickletools.optimize()`: while increasing persistence times dramatically (2.5s to 15.4s), it decreases load times from 4.0s to 1.8s using CPython 3.11.

Is this a worthwhile tradeoff? If cached builds are generated as a one-off batch-job, probably yes. If caching is generated on each build, absolutely not. Optimization can additionally be run offline as a separate job.

How we handle this and fine-tune what we persist and unpersist can be handled in the subsequent task.

# Laptop Timings

## CPython 3.11
Persist: 15.4s
Unpersist: 1.8s

## Pyston 3.8
Persist: 10.3s
Unpersist: 1.7s